Move extensions/pkg/controller/healthcheck/config to extensions/pkg/apis/config
