+++ extensions/pkg/controller/csimigration/reconciler_test.go
@@ -37,7 +37,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
-	healthcheckconfig "github.com/gardener/gardener/extensions/pkg/controller/healthcheck/config"
+	extensionsconfig "github.com/gardener/gardener/extensions/pkg/apis/config"
 )
 
 var _ = Describe("reconciler", func() {
@@ -148,7 +148,7 @@ var _ = Describe("reconciler", func() {
 
 				oldNewClientForShoot := NewClientForShoot
 				defer func() { NewClientForShoot = oldNewClientForShoot }()
-				NewClientForShoot = func(_ context.Context, _ client.Client, _ string, _ client.Options, _ healthcheckconfig.RESTOptions) (*rest.Config, client.Client, error) {
+				NewClientForShoot = func(_ context.Context, _ client.Client, _ string, _ client.Options, _ extensionsconfig.RESTOptions) (*rest.Config, client.Client, error) {
 					return nil, shootClient, nil
 				}
 
@@ -178,7 +178,7 @@ var _ = Describe("reconciler", func() {
 
 				oldNewClientForShoot := NewClientForShoot
 				defer func() { NewClientForShoot = oldNewClientForShoot }()
-				NewClientForShoot = func(_ context.Context, _ client.Client, _ string, _ client.Options, _ healthcheckconfig.RESTOptions) (*rest.Config, client.Client, error) {
+				NewClientForShoot = func(_ context.Context, _ client.Client, _ string, _ client.Options, _ extensionsconfig.RESTOptions) (*rest.Config, client.Client, error) {
 					return nil, shootClient, nil
 				}
